backend/requirements.txt
flask
flask_cors
groq
google.generativeai
python-dotenv